+++ EBO/android/app/src/main/AndroidManifest.xml
@@ -9,6 +9,7 @@
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
     <uses-permission android:name="com.google.android.c2dm.permission.RECEIVE" />
 
+
     <uses-feature android:name="android.hardware.camera" android:required="false" />
 
     <queries>
