+++ EBO/package.json
@@ -35,6 +35,7 @@
     "react-native-progress": "^5.0.1",
     "react-native-safe-area-context": "^4.9.0",
     "react-native-screens": "^3.29.0",
+    "react-native-sound": "^0.11.2",
     "react-native-vector-icons": "^10.0.3",
     "react-native-webview": "^13.8.1"
   },
